refactor(TotalPlan): totalBudget 자료형 변경

+++ src/main/java/com/arom/yeojung/object/TotalPlan.java
@@ -26,7 +26,7 @@ public class TotalPlan {
 
   private LocalDate endDate;
 
-  private BigInteger totalBudget;
+  private Long totalBudget;
 
   private String totalPlanDescription;
